authentication completed

+++ restapi25/01-backend-fetching-list-of-posts/controllers/feed.js
@@ -3,6 +3,7 @@ const fs = require('fs');
 const path = require('path');
 
 const Post = require('../models/post');
+const User = require('../models/user');
 
 exports.getPosts = (req, res, next) => {
   const currentPage = req.query.page || 1;
@@ -50,28 +51,34 @@ exports.createPost = (req, res, next) => {
   const content = req.body.content;
   // Create post in db
 
+  let creator;
   const post = new Post({
-
     title: title,
     content: content,
     imageUrl: imageUrl,
-    creator: { name: 'yasir' }
+    creator: req.userId
   });
-  post.save().then((result) => {
-
-    // console.log('result ', result);
-    res.status(201).json({
-      message: 'Post Created Successfully',
-      post: result
+  post.save()
+    .then((result) => {
+      return User.findById(req.userId);
     })
-
-  }).catch((err) => {
-
-    if (!err.statusCode) {
-      err.statusCode = 500;
-    }
-    next(err);
-  });
+    .then((user) => {
+      creator = user;
+      user.posts.push(post);
+      return user.save();
+    })
+    .then((result) => {
+      res.status(201).json({
+        message: 'Post Created Successfully',
+        post: post,
+        creator: { _id: creator._id, name: creator.name }
+      })
+    }).catch((err) => {
+      if (!err.statusCode) {
+        err.statusCode = 500;
+      }
+      next(err);
+    });
 };
 
 exports.getPost = (req, res, next) => {
@@ -104,7 +111,6 @@ exports.updatePost = (req, res, next) => {
     error.statusCode = 422;
     throw error;
   }
-
   const postId = req.params.postId;
   const title = req.body.title;
   const content = req.body.content;
@@ -113,7 +119,6 @@ exports.updatePost = (req, res, next) => {
   if (req.file) {
     imageUrl = req.file.path.replace("\\", "/");
   };
-
 
   if (!imageUrl) {
     const error = new Error('No file picked');
@@ -126,6 +131,11 @@ exports.updatePost = (req, res, next) => {
       if (!post) {
         const error = new Error('Could not find product');
         error.statusCode = 404;
+        throw error;
+      }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not Authorized');
+        error.statusCode = 403;
         throw error;
       }
       if (imageUrl !== post.imageUrl) {
@@ -161,11 +171,23 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not Authorized');
+        error.statusCode = 403;
+        throw error;
+      }
       clearImage(post.imageUrl);
       return Post.findByIdAndRemove(postId);
     })
     .then((result) => {
-      console.log(result);
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      // console.log(result);
+      user.posts.pull(postId);
+      return user.save();
+    })
+    .then((result) => {
       res.status(200).json({
         message: 'Post Deleted Successfully',
       });
